+++ HISTORY.md
@@ -90,6 +90,7 @@
 * updated "animation" library (purged unused blocks)
 * updated "audioComp" library (purged unused blocks)
 * updated "frequency distribution" library (changed PIPE to use EMPTY prim)
+* Maps: added support for zxy, zyx and xyz style maps and satellite imagery
 
 ### 2019-05-29
 * Threads, Objects: added "Find First" primitive to lists category